ajout des fonctions et classes pour la creation de Categories Tags Types et Ressources La creation de ressources n'est pas encore fonctionnelle

+++ src/main/java/org/acme/controller/UtilisateursController.java
@@ -51,7 +51,7 @@ public class UtilisateursController {
     @GET
     @Path("/{id}")
     @RolesAllowed("Utilisateur")
-    public Response getUser(@PathParam("id") Long id) {
+    public Response getUser(@PathParam("id") int id) {
         Utilisateurs user = utilisateurService.findById(id);
         if (user != null) {
             return Response.ok(user).build();
